Day  #3 Learning Laravel
i learned a lot of this like middleware use for both routing && static function in the PostControllers, revised the association of a user to a model (like Posts, messages, tweets etc..) and alot of this as well

+++ app/Http/Controllers/PostController.php
@@ -4,10 +4,14 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
-use function Pest\Laravel\post;
-
-class PostController extends Controller
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
+class PostController extends Controller implements HasMiddleware
 {
+    public static function middleware() {
+        new Middleware('auth:sanctum', except:  ["index", "show"]);
+    }
     /**
      * Display a listing of the resource.
      */
@@ -33,7 +37,9 @@ class PostController extends Controller
             'title' => 'required|max:255',
             'body' => 'required',
         ]);
-       $post = Post::create($validated);
+        // creates the post from the user.
+        $post = $request->user()->posts()->create($validated);
+
         return ['post'=>$post];
     }
 
@@ -58,6 +64,7 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        Gate::authorize('modify', $post);
         $validate = $request->validate([
             'title' => 'required|max:255',
             'body' => 'required'
@@ -73,11 +80,13 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        Gate::authorize('modify', $post);
+        // message for displaying what title is removed;
+        $message = $post->title . ": is deleted";
         // no need to give the id (since we will get it from the url. (unsafe)).
         $post->delete();
-        // message for displaying what title is removed;
-        $message = $post->title . ": is gone";
 
-        return ['message' =>  $message];
+        return ['Success' =>  $message];
     }
 
+}
